+++ packages/terra-date-picker/tests/jest/DateInput.test.jsx
@@ -4,7 +4,7 @@ import { mountWithIntl, renderWithIntl } from 'terra-enzyme-intl';
 import ThemeContextProvider from 'terra-theme-context/lib/ThemeContextProvider';
 import DateInput from '../../lib/DateInput';
 
-jest.mock('uuid/v4', () => () => '00000000-0000-0000-0000-000000000000');
+jest.mock('uuid', () => ({ v4: () => '00000000-0000-0000-0000-000000000000' }));
 
 it('should render a default date input', () => {
   const dateInput = renderWithIntl(<DateInput />);