+++ packages/terra-slide-panel/src/SlidePanel.jsx
@@ -5,7 +5,7 @@ import classNamesBind from 'classnames/bind';
 import ThemeContext from 'terra-theme-context';
 import VisuallyHiddenText from 'terra-visually-hidden-text';
 import { injectIntl } from 'react-intl';
-import uuidv4 from 'uuid/v4';
+import { v4 as uuidv4 } from 'uuid';
 import styles from './SlidePanel.module.scss';
 
 const cx = classNamesBind.bind(styles);