+++ packages/terra-worklist-data-grid/src/WorklistDataGrid.jsx
@@ -1,4 +1,6 @@
-import React, { useContext, useRef, useCallback } from 'react';
+import React, {
+  useContext, useRef, useCallback,
+} from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import ThemeContext from 'terra-theme-context';
@@ -124,6 +126,7 @@ function WorklistDataGrid(props) {
       default:
         return;
     }
+
     if (nextRow >= grid.current.rows.length || nextCol >= grid.current.rows[0].cells.length) {
       event.preventDefault();
       return;
@@ -150,8 +153,9 @@ function WorklistDataGrid(props) {
     // Determine which cell was clicked. In the event that the user holds the mouse across multiple cells,
     // the originating cell is the clicked cell/active element.
     const clickedCell = event.target.closest('td,th') || document.activeElement.closest('td,th');
+
+    // Ensure that the cell is focusable and is a valid cell
     if (!clickedCell) {
-      // If anything other than a table data or table header cell is clicked, ignore the click.
       return;
     }
     // Remove Tab stop from previous cell in table that has focus and set it to the cell that was clicked.
@@ -169,19 +173,36 @@ function WorklistDataGrid(props) {
 
     // Determine whether cell is a header or grid cell
     const WorklistCellTag = props.rowHeaderIndex === cellColumnIndex ? 'th' : 'td';
+    let cellAriaLabel;
+
+    if (cell.isMasked) {
+      cellAriaLabel = intl.formatMessage({ id: 'Terra.worklistDataGrid.maskedCell' });
+    } else if (!cell.content) {
+      cellAriaLabel = intl.formatMessage({ id: 'Terra.worklistDataGrid.blank' });
+    }
 
     return (
       // Return worklist data grid cell component
       <WorklistCellTag
         key={cellColumnIndex}
         {...tabIndex}
-        className={cx('worklist-data-grid-row-header', { masked: cell.isMasked })}
-        aria-label={cell.isMasked ? intl.formatMessage({ id: 'Terra.worklistDataGrid.maskedCell' }) : undefined}
+        className={cx('worklist-data-grid-cell', { masked: cell.isMasked, selectable: !(cell.isMasked || cell.isSelectable === false), blank: !cell.content })}
+        aria-label={cellAriaLabel}
       >
-        <div className={cx('cell-content')}>{cell.content}</div>
+        {!cell.isMasked && cell.content
+          && <div className={cx('cell-content')}>{cell.content}</div>}
       </WorklistCellTag>
     );
   };
+
+  // const buildColumn = (column) => {
+  //   const width = column.width || props.columnWidth;
+  //   const height = props.columnHeaderHeight;
+  //   return (
+  //     /* eslint-disable react/forbid-dom-props */
+  //     <th key={column.id} className={cx('worklist-data-grid-column-header', { selectable: !(column.isSelectable === false) })} tabIndex="-1" style={{ width, height }}>{column.displayName}</th>
+  //   );
+  // };
 
   const buildColumn = (column) => (
     <HeaderCell
@@ -195,21 +216,25 @@ function WorklistDataGrid(props) {
   const buildColumns = (allColumns) => {
     if (allColumns?.length > 0) {
       return (
-        <tr>
-          {allColumns.map(columnData => (buildColumn(columnData)))}
+        <tr height={props.columnHeaderHeight}>
+          {allColumns.map(column => (buildColumn(column)))}
         </tr>
       );
     }
     return undefined;
   };
 
-  const buildRow = (row) => (
-    <tr key={row.id} className={cx('worklist-data-grid-row')}>
-      {row.cells.map((cell, cellColumnIndex) => (
-        getCellData(cell, cellColumnIndex)
-      ))}
-    </tr>
-  );
+  const buildRow = (row) => {
+    const height = props.rowHeight;
+    return (
+      // eslint-disable-next-line react/forbid-dom-props
+      <tr key={row.id} className={cx('worklist-data-grid-row')} style={{ height }}>
+        {row.cells.map((cell, cellColumnIndex) => (
+          getCellData(cell, cellColumnIndex)
+        ))}
+      </tr>
+    );
+  };
 
   const buildRows = (allRows) => (
     allRows.map((row) => (buildRow(row)))