+++ packages/terra-application-navigation/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 1.67.0 - (June 5, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 1.66.0 - (May 11, 2023)
 
 * Changed