+++ packages/terra-tabs/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## 6.70.0 - (June 12, 2023)
+
+* Changed
+  * Locked `uuid` version to `8.2.0`.
+
+## 6.69.0 - (June 5, 2023)
+
+* Changed
+  * Updated `uuid` to `v8` for consistency with other components.
+
 ## 6.68.0 - (May 11, 2023)
 
 * Changed