+++ packages/terra-worklist-data-grid/CHANGELOG.md
@@ -1,7 +1,13 @@
 # Changelog
 
 ## Unreleased
+* Changed
+  * Removed ability to set variable row heights
+  * Updated styling on cell focus.
 * Added
   * Added masked cell support.
+  * Added hover state for selectable cells.
+  * Added blank cell support.
 
-* Initial stable release
+## 0.1
+  * Initial stable release.